Token Burn Precompile Refactoring (#700)
* introduces TokenBurnLogic


* introduces TokenBurnLogic


* applies patch from TokenMintTransitionLogic refactoring


* minor changes


* aligns burn with mint operation


* minor fixes


* adds missing license


* adds missing annotation


* injects BurnLogic in TokenBurnTransitionLogic
